requirements.txt
fastapi
uvicorn
ollama
python-dotenv
requests